+++ examples/python/utils/utils.py
@@ -25,7 +25,7 @@ def center_crop(img, dim):
         dim (tuple): Shape of the output image
 
     Returns:
-        [numpy.ndarray]: Image croppepd to dim shape
+        [numpy.ndarray]: Image cropped to dim shape
     """
 
     width, height = img.shape[1], img.shape[0]
@@ -127,9 +127,12 @@ def postprocess(response, k=5):
         [numpy.ndarray]: topK values
     """
     outputs = response.getOutputs()
-    response_data = outputs[0].getFp32Data()
-    response_data = np.argsort(response_data)
-    return response_data[-k:][::-1]
+    responses = []
+    for output in outputs:
+        response_data = output.getFp32Data()
+        response_data = np.argsort(response_data)
+        responses.append(response_data[-k:][::-1])
+    return responses
 
 
 def preprocess_pt(image_location, input_size):