+++ tests/workers/test_migraphx.py
@@ -18,38 +18,74 @@ import sys
 import pytest
 import numpy as np
 import proteus
+import cv2
 
 from helper import run_benchmark, root_path
 
 sys.path.insert(0, os.path.join(root_path, "examples/python"))
-from utils.utils import preprocess, postprocess
+from utils.utils import postprocess
+
+
+# The make_nxn and preprocess functions are based on an migraphx example at
+# AMDMIGraphx/examples/vision/python_resnet50/resnet50_inference.ipynb
+# The mean and standard dev. values used for this normalization are requirements of the Resnet50 model.
+
+def make_nxn(image, n):
+    """ 
+    Crop an image to square and then resize to desired dimension n x n
+    """
+    width = image.shape[1]
+    height = image.shape[0]
+    if height > width:
+        dif = height - width
+        bar = dif // 2
+        square = image[(bar + (dif % 2)) : (height - bar), :]
+        return cv2.resize(square, (n, n))
+    elif width > height:
+        dif = width - height
+        bar = dif // 2
+        square = image[:, (bar + (dif % 2)) : (width - bar)]
+        return cv2.resize(square, (n, n))
+    else:
+        return cv2.resize(image, (n, n))
+
+def preprocess(img_data):
+    """
+    Normalize array values to the data type, mean and std. dev. required by Resnet50
+
+    Args:
+        img_data (np.array): array in 3 dimensions [channels, rows, cols] with value range 0-255
+
+    Returns:
+        Response: np.array
+    """
+    mean_vec = np.array([0.485, 0.456, 0.406])
+    stddev_vec = np.array([0.229, 0.224, 0.225])
+    norm_img_data = np.zeros(img_data.shape).astype("float32")
+    for i in range(img_data.shape[0]):
+        norm_img_data[i, :, :] = (img_data[i, :, :] / 255 - mean_vec[i]) / stddev_vec[i]
+    return norm_img_data
 
 
 @pytest.fixture(scope="class")
 def model_fixture():
-    return "TfZendnn"
+    return "Migraphx"
 
 
 @pytest.fixture(scope="class")
 def parameters_fixture():
     return {
         "model": str(
-            root_path / "external/tensorflow_models/resnet_v1_50_baseline_6.96B_922.pb"
-        ),
-        "input_node": "input",
-        "output_node": "resnet_v1_50/predictions/Reshape_1",
-        "input_size": 224,
-        "output_classes": 1000,
-        "inter_op": 64,
-        "intra_op": 1,
+            root_path / "external/artifacts/migraphx/resnet50v2/resnet50-v2-7.onnx"
+        )
     }
 
 
-@pytest.mark.extensions(["tfzendnn"])
+@pytest.mark.extensions(["migraphx"])
 @pytest.mark.usefixtures("load")
-class TestTfZendnn:
+class TestMigraphx:
     """
-    Test the TfZendnn worker
+    Test the Migraphx worker
     """
 
     def send_request(self, request, check_asserts=True):
@@ -77,7 +113,7 @@ class TestTfZendnn:
         if check_asserts:
             assert not response.isError(), response.getError()
             assert response.id == ""
-            assert response.model == "TFModel"
+            assert response.model == "migraphx"
             outputs = response.getOutputs()
             assert len(outputs) == num_inputs
             for index, output in enumerate(outputs):
@@ -87,33 +123,44 @@ class TestTfZendnn:
         return response
 
     @pytest.mark.parametrize("num", [1, 8])
-    def test_tfzendnn_0(self, num):
+    def test_migraphx(self, num):
         """
-        Send a request to tf model as tensor data
+        Send a request to model as tensor data
         """
-        image_path = str(root_path / "tests/assets/dog-3619020_640.jpg")
+        image_paths = [ 
+            str(root_path / "tests/assets/dog-3619020_640.jpg"), 
+            str(root_path / "tests/assets/bicycle-384566_640.jpg")
+        ]
 
-        preprocessing = {"input_size": 224, "resize_method": "crop"}
+        gold_responses = [
+            [259, 261, 157, 154, 230],
+            [671, 444, 518, 665, 638]
+        ]
+
+        assert len(image_paths) == len(gold_responses)
+        image_num = len(image_paths)
 
         batch = num
         images = []
-        for _ in range(batch):
-            images.append(
-                preprocess(
-                    image_path,
-                    input_size=preprocessing["input_size"],
-                    resize_method=preprocessing["resize_method"],
-                )
-            )
+        for i in range(batch):
+            # Load a picture
+            img = cv2.imread(image_paths[i % image_num]).astype("float32")
+            # Crop to a square, resize
+            img = make_nxn(img, 224)
+            #  Normalize contents with values specific to Resnet50
+            img = img.transpose(2, 0, 1)
+            img = preprocess(img)
+            images.append(img)
         request = proteus.ImageInferenceRequest(images, True)
         response = self.send_request(request)
         top_k_responses = postprocess(response, 5)
-        gold_response_output = [259, 261, 154, 260, 263]
-        for top_k in top_k_responses:
-            assert (top_k == gold_response_output).all()
 
-    @pytest.mark.benchmark(group="TfZendnn")
-    def test_benchmark_tfzendnn(self, benchmark, model_fixture, parameters_fixture):
+        for i, top_k in enumerate(top_k_responses):
+            # Look for the top 5 categories (not underlying scores)
+            assert (top_k == gold_responses[i % image_num]).all()
+
+    @pytest.mark.benchmark(group="Migraphx")
+    def test_benchmark_Migraphx(self, benchmark, model_fixture, parameters_fixture):
 
         batch_size = 16
         input_size = parameters_fixture.get("input_size")
@@ -131,5 +178,5 @@ class TestTfZendnn:
             "config": "N/A",
         }
         run_benchmark(
-            benchmark, "TfZendnn", self.rest_client.modelInfer, request, **options
+            benchmark, "Migraphx", self.rest_client.modelInfer, request, **options
         )